+++ examples/flask_server_decorators.py
@@ -5,9 +5,9 @@ from flask import Flask, jsonify, make_response
 from examples.grpc_resources.grpc_demo_pb2 import DemoMessage
 from examples.grpc_resources.grpc_demo_pb2_grpc import DemoServiceStub
 from examples.grpc_server import grpc_port
-from stackdriver_logging.flask_helpers.decorators import trace_and_log_route, trace_and_log_exception
-from stackdriver_logging.jsonlog import get_logger
-from stackdriver_logging.tracing import generate_new_traced_subspan_values
+from logtrace.flask_helpers.decorators import trace_and_log_route, trace_and_log_exception
+from logtrace.jsonlog import get_logger
+from logtrace.tracing import generate_new_traced_subspan_values
 
 # flask
 app = Flask('demoFlaskLoggerDecorators')
