+++ logtrace/_global_vars.py
