+++ logtrace/flask_helpers/callbacks.py
@@ -1,7 +1,7 @@
 from flask import request
 
-from stackdriver_logging.jsonlog import get_logger
-from stackdriver_logging.tracing import start_traced_span, end_traced_span
+from logtrace.jsonlog import get_logger
+from logtrace.tracing import start_traced_span, end_traced_span
 
 
 def start_span_and_log_request_before(excluded_routes=None, excluded_routes_partial=None):
