+++ logtrace/grpc_helpers/decorators.py
@@ -1,8 +1,8 @@
 import copy
 import functools
 
-from stackdriver_logging.jsonlog import get_logger
-from stackdriver_logging.tracing import start_traced_span, end_traced_span
+from logtrace.jsonlog import get_logger
+from logtrace.tracing import start_traced_span, end_traced_span
 
 
 def _rsetattr(obj, attr, val):
