+++ tests/test_tracing.py
@@ -1,6 +1,6 @@
 from unittest.mock import patch
 
-from stackdriver_logging.tracing import _truncate_str, thread_memory, start_span, SubSpan
+from logtrace.tracing import _truncate_str, thread_memory, start_span, SubSpan
 
 
 def test_truncate_str():
@@ -13,8 +13,8 @@ def test_truncate_str():
     assert trunc_obj == {'value': 'kindoflong', 'truncated_byte_count': 6}
 
 
-@patch('stackdriver_logging.tracing.b3.SubSpan')
-@patch('stackdriver_logging.tracing.b3.start_span')
+@patch('logtrace.tracing.b3.SubSpan')
+@patch('logtrace.tracing.b3.start_span')
 def test_g_child_span_count(startspan, subspan):
     start_span({}, '', '', '')
     assert thread_memory.child_span_count == 0
